requirements.txt
fastapi
uvicorn
python-multipart
requests
streamlit
pydantic
chromadb
beautifulsoup4
sentence-transformers
google-generativeai
python-dotenv
unstructured
pdfplumber
PyMuPDF
protobuf==3.20.
huggingface-hub
